requirements.txt
numpy~=1.19.5
tensorflow-gpu~=2.4.1
gym~=0.18.0
pandas~=1.2.3
cffi~=1.14.5
ipython~=7.22.0
pip~=20.0.2
Keras~=2.4.3
Cython~=0.29.22
setuptools~=56.0.0
scipy~=1.6.2
wheel~=0.36.2
notebook~=6.3.0
jupyterlab~=3.0.14
tensorboard~=2.4.1
